Rename misc/utils.py -> utils/misc.py
