+++ softlearning/utils/misc.py
@@ -1,5 +1,4 @@
 import collections
-import datetime
 import os
 import random
 
@@ -21,24 +20,6 @@ def set_seed(seed):
     np.random.seed(seed)
     tf.compat.v1.set_random_seed(seed)
     print("Using seed {}".format(seed))
-
-
-def datetimestamp(divider='-', datetime_divider='T'):
-    now = datetime.datetime.now()
-    return now.strftime(
-        '%Y{d}%m{d}%dT%H{d}%M{d}%S'
-        ''.format(d=divider, dtd=datetime_divider))
-
-
-def datestamp(divider='-'):
-    return datetime.date.today().isoformat().replace('-', divider)
-
-
-def timestamp(divider='-'):
-    now = datetime.datetime.now()
-    time_now = datetime.datetime.time(now)
-    return time_now.strftime(
-        '%H{d}%M{d}%S'.format(d=divider))
 
 
 def deep_update(d, *us):
